Add new implementation algorithm
